refactor(cli): extract shared run summary formatter from cli-run and cli-skill
