+++ src/cli-run.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 import type { appConfig as appConfigType } from "./app-config";
-import { formatCompactNumber } from "./chat-format";
 import type { createMessage as createMessageType } from "./chat-session";
 import type { attachFileToSession as attachFileToSessionType } from "./cli-chat";
+import { formatRunSummary } from "./cli-format";
 import type { handlePrompt as handlePromptType } from "./cli-prompt";
 import type { createClient as createClientType } from "./client-factory";
 import type { readResolvedConfigSync as readResolvedConfigSyncType } from "./config";
@@ -156,21 +156,8 @@ export async function runMode(args: string[], deps: RunModeDeps): Promise<void> 
   });
   const durationMs = Date.now() - startMs;
 
-  const totals = session.tokenUsage.reduce(
-    (acc, e) => ({
-      input: acc.input + e.usage.inputTokens,
-      output: acc.output + e.usage.outputTokens,
-      total: acc.total + e.usage.totalTokens,
-      modelCalls: acc.modelCalls + (e.modelCalls ?? 1),
-    }),
-    { input: 0, output: 0, total: 0, modelCalls: 0 },
-  );
-  const durationSec = (durationMs / 1000).toFixed(1);
-  if (totals.total > 0) {
-    printDim(
-      `run: ${durationSec}s, ${formatCompactNumber(totals.total)} tokens (input ${formatCompactNumber(totals.input)}, output ${formatCompactNumber(totals.output)}), ${totals.modelCalls} model calls, ${session.tokenUsage.length} turns`,
-    );
-  }
+  const summary = formatRunSummary("run", session.tokenUsage, durationMs);
+  if (summary) printDim(summary);
 
   if (!success) {
     process.exitCode = 1;
