+++ src/cli-skill.ts
@@ -1,6 +1,7 @@
 import type { appConfig as appConfigType } from "./app-config";
 import type { createMessage as createMessageType } from "./chat-session";
 import type { attachFileToSession as attachFileToSessionType } from "./cli-chat";
+import { formatRunSummary } from "./cli-format";
 import type { handlePrompt as handlePromptType } from "./cli-prompt";
 import type { createClient as createClientType } from "./client-factory";
 import type { readResolvedConfigSync as readResolvedConfigSyncType } from "./config";
@@ -168,21 +169,8 @@ export async function skillMode(args: string[], deps: SkillModeDeps): Promise<vo
   });
   const durationMs = Date.now() - startMs;
 
-  const totals = session.tokenUsage.reduce(
-    (acc, e) => ({
-      input: acc.input + e.usage.inputTokens,
-      output: acc.output + e.usage.outputTokens,
-      total: acc.total + e.usage.totalTokens,
-      modelCalls: acc.modelCalls + (e.modelCalls ?? 1),
-    }),
-    { input: 0, output: 0, total: 0, modelCalls: 0 },
-  );
-  const durationSec = (durationMs / 1000).toFixed(1);
-  if (totals.total > 0) {
-    printDim(
-      `skill: ${durationSec}s, ${totals.total} tokens (input ${totals.input}, output ${totals.output}), ${totals.modelCalls} model calls, ${session.tokenUsage.length} turns`,
-    );
-  }
+  const summary = formatRunSummary("skill", session.tokenUsage, durationMs);
+  if (summary) printDim(summary);
 
   if (!success) {
     process.exitCode = 1;
